Added `ser/arrays` module
