+++ crates/bevy_reflect/src/serde/ser/serializer.rs
@@ -1,3 +1,4 @@
+use crate::serde::ser::arrays::ArraySerializer;
 use crate::serde::ser::enums::EnumSerializer;
 use crate::serde::ser::lists::ListSerializer;
 use crate::serde::ser::maps::MapSerializer;
@@ -5,7 +6,7 @@ use crate::serde::ser::sets::SetSerializer;
 use crate::serde::ser::structs::StructSerializer;
 use crate::serde::ser::tuple_structs::TupleStructSerializer;
 use crate::serde::ser::tuples::TupleSerializer;
-use crate::serde::{ArraySerializer, Serializable};
+use crate::serde::Serializable;
 use crate::{PartialReflect, ReflectRef, TypeRegistry};
 use serde::ser::{Error, SerializeMap};
 use serde::Serialize;
@@ -160,11 +161,9 @@ impl<'a> Serialize for TypedReflectSerializer<'a> {
             ReflectRef::List(value) => {
                 ListSerializer::new(value, self.registry).serialize(serializer)
             }
-            ReflectRef::Array(value) => ArraySerializer {
-                array: value,
-                registry: self.registry,
+            ReflectRef::Array(value) => {
+                ArraySerializer::new(value, self.registry).serialize(serializer)
             }
-            .serialize(serializer),
             ReflectRef::Map(value) => {
                 MapSerializer::new(value, self.registry).serialize(serializer)
             }
